requirements.txt
anthropic>=0.25.0
openai>=1.30.0
pinecone-client>=3.2.0
pyodbc>=5.1.0
snowflake-connector-python>=3.10.0
python-dotenv>=1.0.0
pytest>=8.2.0
